Inicio da incersão e correção de unsigned char para unisgned short para o key

+++ textoEditor2/keyBinds.h
@@ -20,6 +20,6 @@
 
 // function
 
-int readKey(void);
+unsigned short readKey(void);
 
 #endif
